+++ integrations/nvidia/tests/__init__.py
@@ -1,3 +1,6 @@
 # SPDX-FileCopyrightText: 2023-present deepset GmbH <info@deepset.ai>
 #
 # SPDX-License-Identifier: Apache-2.0
+from .conftest import MockBackend
+
+__all__ = ["MockBackend"]
